finalized the script to work now! yay!

+++ client/client.lua
@@ -1,22 +1,21 @@
 Citizen.CreateThread(function()
 	while true do
-		Wait(1000)
+		Wait(Config.RefreshTime)
 		local ped = PlayerPedId()
 		local coords = GetEntityCoords(ped)
 		local isinwater = IsPedSwimming(ped)
 		local waterheight = PerformRaycastFromPed(ped)
-		local indanger = isinwater and (waterheight or 0 >= Config.WaterHeight or waterheight == 0.0) and GetVehiclePedIsIn(ped, false) == 0
-        local inAttack = IsEntityPlayingAnim(ped, "creatures@shark@melee@streamed_core@", "attack_player", 3)
+		local indanger = isinwater and (waterheight or 0) >= Config.WaterHeight and GetVehiclePedIsIn(ped, false) == 0
+		local inAttack = IsEntityPlayingAnim(ped, "creatures@shark@melee@streamed_core@", "attack_player", 3)
 		-- Check if the player is in water and the water height is above the threshold
+		print(indanger)
 		if indanger then
-			print("Player is in water.")
 			if DoesEntityExist(shark) then
-				print("Shark already exists.")
 			else
 				shark = SpawnAgressiveShark(coords)
 				Citizen.Wait(Config.TimeUntilAttack * 1000)
 				if indanger then
-					sharkAttack(coords, shark)
+					sharkAttack(shark)
 				else
 					Config_Framework_Notify_Client(Locale("sharkdespawned"))
 					DespawnShark(shark)
@@ -24,7 +23,7 @@ Citizen.CreateThread(function()
 			end
 		else
 			if shark and not inAttack then
-                DespawnShark(shark)
+				DespawnShark(shark)
 			end
 		end
 	end
@@ -47,13 +46,13 @@ function sharkAttack(shark)
 		Wait(1)
 	end
 	local ped = PlayerPedId()
-	TaskPlayAnim(ped, dict2, anim2, 8.0, 1.0, -1, 1, 0, 0, 0, 0)
+	TaskPlayAnim(ped, dict, anim2, 8.0, 1.0, -1, 1, 0, 0, 0, 0)
 	local model = GetHashKey("a_c_sharktiger")
 	RequestModel(model)
 	while not HasModelLoaded(model) do
 		Wait(1)
 	end
-	local coords2 = GetEntityCoords(shark)
+	local coords2 = GetEntityCoords(ped)
 	local scene = NetworkCreateSynchronisedScene(
 		coords2.x,
 		coords2.y,
@@ -74,57 +73,52 @@ function sharkAttack(shark)
 	NetworkStartSynchronisedScene(scene)
 	Wait(0)
 	local localScene = NetworkGetLocalSceneFromNetworkId(scene)
-	Citizen.Wait(1500)
-	DoScreenFadeOut(1500)
+	Citizen.Wait(1000)
+	DoScreenFadeOut(1000)
+	Citizen.Wait(1000)
 	repeat
 		Wait(0)
-
 	until GetSynchronizedScenePhase(localScene) >= 0.20
-    makeBlood()
-    repeat Wait(0) until GetSynchronizedScenePhase(localScene) >= 0.25
+	makeBlood()
+	repeat Wait(0) until GetSynchronizedScenePhase(localScene) >= 0.25
 	ClearPedTasks(shark)
 	TaskPlayAnim(shark, dict, "swim_turn_r", 8.0, 1.0, -1, 1, 0, 0, 0, 0)
 	NetworkStopSynchronisedScene(scene)
-    DespawnShark(shark, true)
+	DespawnShark(shark, true)
 	Citizen.Wait(1000)
-    Config_Framework_SetDead_Client()
+	Config_Framework_SetDead_Client()
 	DoScreenFadeIn(1000)
-
+	attack = true
 end
 
 function makeBlood()
-    dict = "core"
-    name = "blood_shark_attack"
-    coords = GetEntityCoords(PlayerPedId())
-    x,y,z = table.unpack(coords)
-   
-    CreateThread( function()
-    
-        RequestNamedPtfxAsset(dict)
-        -- Wait for the particle dictionary to load.
-        while not HasNamedPtfxAssetLoaded(dict) do
-            Citizen.Wait(0)
-        end
+	-- dict = "core"
+	-- name = "blood_shark_attack"
+	-- coords = GetEntityCoords(PlayerPedId())
+	-- x, y, z = table.unpack(coords)
 
-        local a = 0
-        while a < 4 do
-           
-          
-            StartParticleFxNonLoopedAtCoord(name, x + 5.0, y + 5.0, z + 1.0, 0.0, 0.0, 0.0, 0.2, false, false, false)
-        
-            a = a + 1
-            
-            -- Wait 500ms before triggering the next particle.
-            Citizen.Wait(500)
-        end
-    end)
+	-- CreateThread(function()
+	-- 	RequestNamedPtfxAsset(dict)
+	-- 	-- Wait for the particle dictionary to load.
+	-- 	while not HasNamedPtfxAssetLoaded(dict) do
+	-- 		Citizen.Wait(0)
+	-- 	end
+
+	-- 	local a = 0
+	-- 	while a < 4 do
+	-- 		StartParticleFxNonLoopedAtCoord(name, x, y, z, 0.0, 0.0, 0.0, 0.2, false, false, false)
+
+	-- 		a = a + 1
+
+	-- 		-- Wait 500ms before triggering the next particle.
+	-- 		Citizen.Wait(500)
+	-- 	end
+	-- end)
 end
 
 -- RegisterCommand('testan', function(source, args, rawCommand)
 --     local dict = "creatures@shark@move"
 --     local anim = "attack"
---     local dict2 = "creatures@shark@move"
---     local anim2 = "attack_player"
 
 --     RequestAnimDict(dict)
 --     while not HasAnimDictLoaded(dict) do
@@ -244,13 +238,17 @@ end
 -- end)
 
 function DespawnShark(shark, attack)
-    if attack == nil or false then
-    Config_Framework_Notify_Client(Locale("sharkdespawned"))
-    end
-    TaskGoToCoordAnyMeans(shark, 0.0, 0.0, 0.0, 1.0, 0, false, 786603, 0xbf800000)
-    RemoveBlip(aggresiveblip)
-    Citizen.Wait(2000)
-    DeleteEntity(shark)
+	if DoesEntityExist(shark) == false then
+		return
+	end
+	if attack == nil or false then
+		Config_Framework_Notify_Client(Locale("sharkdespawned"))
+	end
+	TaskGoToCoordAnyMeans(shark, 0.0, 0.0, 0.0, 1.0, 0, false, 786603, 0xbf800000)
+	RemoveBlip(aggresiveblip)
+	Citizen.Wait(2000)
+	DeleteEntity(shark)
+	attack = nil
 end
 
 function SpawnAgressiveShark(coords)
@@ -259,38 +257,37 @@ function SpawnAgressiveShark(coords)
 	while not HasModelLoaded(model) do
 		Wait(1)
 	end
-	shark = CreatePed(4, model, coords.x, coords.y, coords.z - 5, 0.0, true, true)
-	Config.Functions.Notify()
+	shark = CreatePed(4, model, coords.x + 5, coords.y + 5, coords.z - 5, 0.0, true, true)
+	Config_Framework_Notify_Client(Locale("sharkspawned"))
 	aggresiveblip = AddBlipForEntity(shark)
-    SetBlipSprite(aggresiveblip, 68)
-    SetBlipColour(aggresiveblip, 1)
-    SetBlipScale(aggresiveblip, 0.8)
-    SetBlipAsShortRange(aggresiveblip, false)
-    BeginTextCommandSetBlipName("STRING")
-    AddTextComponentString(Locale("sharkname"))
-    EndTextCommandSetBlipName(aggresiveblip)
-    TaskFollowToOffsetOfEntity(shark, PlayerPedId(), 0.0, 0.0, 0.0, 1.0, -1, 1.0, true)
+	SetBlipSprite(aggresiveblip, 68)
+	SetBlipColour(aggresiveblip, 1)
+	SetBlipScale(aggresiveblip, 0.8)
+	SetBlipAsShortRange(aggresiveblip, false)
+	BeginTextCommandSetBlipName("STRING")
+	AddTextComponentString(Locale("sharkname"))
+	EndTextCommandSetBlipName(aggresiveblip)
+	TaskFollowToOffsetOfEntity(shark, PlayerPedId(), 5.0, 5.0, 0.0, 1, -1, 0.0, true)
 	return shark
 end
 
 -- Function to perform a raycast from a ped downward and return the Z coordinate if it hits a surface
 function PerformRaycastFromPed(ped)
-    local playerPos = GetEntityCoords(ped, true)
-    local startPoint = vector3(playerPos.x, playerPos.y, playerPos.z) -- Adjust the starting height as needed
-    local endPoint = vector3(playerPos.x, playerPos.y, playerPos.z - 1000000.0) -- A very large negative value
+	local playerPos = GetEntityCoords(ped, true)
+	local startPoint = vector3(playerPos.x, playerPos.y, playerPos.z)    -- Adjust the starting height as needed
+	local endPoint = vector3(playerPos.x, playerPos.y, playerPos.z - 250.0) -- A very large negative value
 
-    local rayHandle = StartShapeTestRay(startPoint.x, startPoint.y, startPoint.z, endPoint.x, endPoint.y, endPoint.z, -1, ped, 0)
-    local _, hit, hitCoords, _, _ = GetShapeTestResult(rayHandle)
-
-    if hit then
-        -- A surface was hit
-        return math.abs(hitCoords.z)
-    else
-        -- Raycast did not hit anything
-        return 0
-    end
+	local rayHandle = StartShapeTestRay(startPoint.x, startPoint.y, startPoint.z, endPoint.x, endPoint.y, endPoint.z, -1,
+		ped, 0)
+	local _, hit, hitCoords, _, _ = GetShapeTestResult(rayHandle)
+	zone = GetNameOfZone(playerPos.x, playerPos.y, playerPos.z)
+	if hit ~= 0 then
+		-- A surface was hit
+		return math.abs(hitCoords.z)
+	elseif zone == "OCEANA" then
+		return 300.0
+	end
 end
-
 
 function Locale(msg)
 	return Translation[Config.Locale][msg]
